started- and terminated-messages, etc
Add a `{name} started` message to make it easier to get statistics on what recipes are being used.
Maybe, still need to see if this is easier to analyze than the transactions/spans.

Log an error if the transaction is not manually consumed to make sure any such cases are caught.
Prompted by me being unsure if interrupting `just` would propagate as an error or just drop everything.

Handle the signal error differently to hopefully make it possible to filter out some recipes that run indefinitely from statistics.
The span status can be filtered in explorer, but not in dashboard.
Setting a tag on the transaction doesn't appear to help for the dashboard either.

Capture details about where a failure occurred in a debug message mostly to see if the recipe is different than the name.